Separate ODF spec from UFP specific changes
